requirements.txt
celery==4.3.0
envparse==0.2.0
grpcio==1.21.1
grpcio-tools==1.21.1
playsound==1.2.2
requests==2.22.0
vext==0.7.3
vext.gi==0.7.0